add controllers, remove spring security 6, it works

+++ backend/contactapi/src/main/java/com/abbasza/contactapi/service/ContactService.java
@@ -19,36 +19,34 @@ import java.util.UUID;
 @Slf4j
 public class ContactService {
     private final ContactRepo contactRepo;
-    private UserService userService;
+    private final UserService userService;
 
     @Autowired
-    public ContactService(ContactRepo contactRepo) {
+    public ContactService(ContactRepo contactRepo, UserService userService) {
         this.contactRepo = contactRepo;
-    }
-
-    public void setUserService(UserService userService) {
         this.userService = userService;
     }
 
     @Value("${DB_DATABASE}")
     private String valuee;
 
-    public Page<Contact> getAllContacts(int page, int size){
+
+    public Page<Contact> getAllContacts(int page, int size) {
         return contactRepo.findAll(PageRequest.of(page, size, Sort.by("firstName")));
     }
 
-    public Contact getContactById(UUID id){
+    public Contact getContactById(UUID id) {
         return contactRepo.findById(id).orElseThrow(() -> {
-            log.error("Contact: {} Not Found", id);
+            log.info("Contact: {} Not Found", id);
             return new RuntimeException("Contact Not Found" + id);
         });
     }
 
-    public Contact saveContact(Contact contact, String userName){
-        try{
-            log.info("Creating Contact: {}", contact.getId());
+    public Contact saveContact(Contact contact, String email) {
+        try {
             Contact savedContact = contactRepo.save(contact);
-            User user = userService.findUserByUserName(userName);
+            log.info("Creating Contact: {}", savedContact.getId());
+            User user = userService.findUserByEmail(email);
             user.getContacts().add(savedContact);
             userService.saveUser(user);
             return savedContact;
@@ -58,7 +56,31 @@ public class ContactService {
         }
     }
 
-    public void deleteContactById(UUID id){
-        contactRepo.deleteById(id);
+    public void saveContact(Contact contact) {
+        try {
+            log.info("Updating Contact: {}", contact.getId());
+            contactRepo.save(contact);
+        } catch (RuntimeException e) {
+            log.error("Error occured while updating Contact: {}", contact.getId());
+            throw new RuntimeException(e);
+        }
+    }
+
+    public boolean deleteContactById(UUID id, String email) {
+        try {
+            log.info("Deleting Contact: {}", id);
+            User user = userService.findUserByEmail(email);
+            boolean removed = user.getContacts().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                contactRepo.deleteById(id);
+            }else{
+                log.info("Contact Not Found: {}", id);
+            }
+            return removed;
+        } catch (Exception e) {
+            log.error("Error occured while deleting Contact: {}", id);
+            throw new RuntimeException(e);
+        }
     }
 }
